ModelPipeline/CloudFunctions/requirement.txt
functions-framework==3.*
mysql-connector-python
google-auth
requests